EOW-334TO DO3 sub-tasksUser Threshold level should remember what the user has, currently does not display what that users threshold is

+++ src/3. Database/Well/dbo/StoredProcedures/CreditThresholdSave.sql
@@ -1,6 +1,6 @@
 ﻿Create PROCEDURE [dbo].[CreditThresholdSave]
 	@ThresholdLevelId INT,
-	@Threshold INT,
+	@Threshold DECIMAL(10,2),
 	@DateCreated DATETIME,
 	@DateUpdated DATETIME,
 	@CreatedBy VARCHAR(50),
